KUDU-3535 Clear log cache while tombstoning a tablet replica.
The log cache of a replica still exists even if the replica has been
already tombstoned. This problem might take place if we decrease
the replication factor of a table with high throughput.

So we should clear the log cache while deleting the replica with
delete type TABLET_DATA_TOMBSTONED.

Change-Id: I6cf545e604f80d41e7ebd9660acfd2e928cd27a9
Reviewed-on: http://gerrit.cloudera.org:8080/20822
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>