Do not enter ALTERING state just for replication factor changed.
If we change the replication factor of a table without any tablets,
the table will stay in ALTERING state since exiting this state needs
tablet reports from the tablets of this table.

for now the condition of entering ALTERING state is as below:
    has_metadata_changes && (table->num_tablets() >
        tablets_to_drop.size() || num_replicas_changed);

Actually even if the number of replicas has been changed, the former
condition is also needed. So rollback the code to:
    has_metadata_changes && table->num_tablets() >
        tablets_to_drop.size();

Change-Id: I1d20da7c0dd5912790aaa46e9fff366b2973d7a4
Reviewed-on: http://gerrit.cloudera.org:8080/20830
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>