[java] add sequenceId into the tracked RPCs' trace
This patch adds sequenceId into the result of KuduRpc.ToString() for
tracked RPCs (i.e. Write RPC as of now).  That's to provide better
tracking of retries performed in the context of exactly-once semantics.
I found this information would be helpful when troubleshooting
particular scenarios.

Otherwise, there are no functional modifications in this changelist.

Change-Id: I55bd45f13b406a5aad7532f3a449a7c670d9f09d
Reviewed-on: http://gerrit.cloudera.org:8080/20857
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Kudu Jenkins