[java] fix NPE in Connection::exceptionCaught()
Before this patch, an NPE might be thrown by the Java Kudu client
when trying to establish a connection to a Kudu server.  I have a test
scenario to reproduce the issue, but it requires a lot of latency and
failure injection on the server side, which I'm not going to add just to
reproduce this issue.  However, the issue is real and might be hit
in field when Kudu servers run on very busy nodes, sharing the node's
CPU and memory resources with many other active processes, and Kudu's
acceptor thread might be often de-scheduled, so a long queue of pending
connections grows at the server's RPC socket.

It's quite clear the code before this patch had the corresponding
deficiency, so I think this patch is good to go without adding a test
that reproducing the issue.  If I can eventually justify adding
all the latency and failure injection at the server side, I'll add the
corresponding test as well in a follow-up patch.

Change-Id: Ibd6eed2ecdcadab7eaff301f3a3f3d5dcacfd83d
Reviewed-on: http://gerrit.cloudera.org:8080/20858
Tested-by: Kudu Jenkins
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>