[rpc] introduce rpc_pending_connections metric
One of the prior patches introduced 'rpc_listen_socket_rx_queue_size'
histogram-type metric.  That one provides information on the history of
the RPC listening sockets' backlog of pending connections, but this new
'rpc_pending_connections' metric reports on the current total number of
pending RPC connections across all the RPC endpoints of a Kudu server.
The newly introduced metric is useful for assessing the current status
of the listening RPC sockets.  It's also useful for post-mortem analysis
using the diagnostic logs that store snapshots of this new metric tied
to the timestamps when they were captured.

Change-Id: I0bfdaf9047f43495df85edba3200286f743330a8
Reviewed-on: http://gerrit.cloudera.org:8080/20949
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>