[test] a small clean-up on StringGaugePrometheusTest
I took a quick look at KUDU-3549, and a natural first step was to check
whether we have a proper test coverage for handling StringGauge in when
dumping metrics in Prometheus format.  The result was this patch, adding
a couple new sub-scenarios into MetricsTest.StringGaugePrometheusTest.

While this didn't lead to discovering the root case of the reported
issue, the new sub-scenarios are still useful since they provide extra
coverage for the related functionality.  I also took the liberty of
cleaning up the code a bit.

Change-Id: Ie495f3bae91facd98e9ce61ee86f34dea8b08087
Reviewed-on: http://gerrit.cloudera.org:8080/20984
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>