KUDU-613: Add SLRU Cache metrics
This patch adds segment-specific metrics for both
the probationary and the protected segment along
with high-level metrics for the entire SLRU cache.
It also adds these same metrics to the block cache
so it can support a cache with SLRU eviction policy.

This patch adds a template parameter to the
SLRUCacheShard class to indicate which segment
a shard belongs to.

Change-Id: I1c8181ec6bea301605aaef7db0003c3eaef3072d
Reviewed-on: http://gerrit.cloudera.org:8080/21389
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>