[Java] Fix concurrent problem while traversing the traces of KuduRpc.
When the RPC request which is sent by java client times out, the
RpcTimeoutTask is called by HashedWheelTimer. During the running
of this task, KuduRpc#toString() is called. This method will traverse
the traces(a synchronizedList) without any protection, which leads to
exception in timeout task and then the callback will not be triggered.

The exception message is as follow:

An exception was thrown by TimerTask.
java.util.ConcurrentModificationException: null
        at java.util.ArrayList$Itr.checkForComodification
        at java.util.ArrayList$Itr.next
        at org.apache.kudu.client.RpcTraceFrame
           .getHumanReadableSummaryStringForTraces
        at org.apache.kudu.client.KuduRpc.toString
        at java.lang.String.valueOf
        at java.lang.StringBuilder.append
        at org.apache.kudu.client.KuduRpc$RpcTimeoutTask.run
        at org.apache.kudu.shaded.io.netty.util
           .HashedWheelTimer$HashedWheelTimeout.expire

So make a copy of the traces and then traverse the copy. This method
is mostly called when some exceptions are raised, so the influence on
performance should be acceptable.

Change-Id: I1a642d93d01ebcfa05b01fe263023b5580d2542b
Reviewed-on: http://gerrit.cloudera.org:8080/21041
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>