[java] methods for setting run-time flags via test harness
When I was trying to reproduce an issue seen in the field, I found
that necessary functionality was missing in the Java test harness.
This patch addresses the deficiency, introducing the corresponding
methods to set run-time flags for Kudu masters and tablet servers.
This patch also updates the test harness class to provides a means
to restart a tablet server that hosts a particular tablet replica.

Change-Id: I5ed12b2ef9fd077534528361f6bb42efe3730182
Reviewed-on: http://gerrit.cloudera.org:8080/21093
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>