[Tool] Return not OK status when copying tablets failed
Currently, 'local_replica copy_from_remote' command does not
return a non-OK status when some tablets copying failed. Therefore
it is not possible to use the return status to know whether it is
failed or succeeded. Only the log can show some useful message.

This patch fixes this problem so now the CLI tool exits with non-OK
status code when some tablets failed to copy.

Change-Id: Ic957cbc379645e0607c1c2a3bc568e20afc126b2
Reviewed-on: http://gerrit.cloudera.org:8080/21089
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>