[thirdparty] upgrade boost up to 1.84.0
Upgrade Boost library up to 1.84.0 version.  At least, this version
properly handles BOOST_NO_CXX98_FUNCTION_BASE macro in
boost/config/stdlib/libcpp.hpp, defining it with CLANG 15.
Otherwise, it would be necessary to add BOOST_NO_CXX98_FUNCTION_BASE
into CFLAGS and CXXFLAGS to avoid compilation errors like the below
at least when compiling with CLANG 15:

  In file included from thirdparty/installed/uninstrumented/include/boost/type_index/stl_type_index.hpp:47:
  thirdparty/installed/uninstrumented/include/boost/container_hash/hash.hpp:131:33: error: no template named 'unary_function' in namespace 'std'; did you mean '__unary_function'?
              struct hash_base : std::unary_function<T, std::size_t> {};
                                 ~~~~~^~~~~~~~~~~~~~
                                      __unary_function

This patch fixes compilation of the mustache library in the 3rd-party
with CLANG 15 which comes with Xcode 15.3 on macOS: std::unary_function
has been deprecated since C++11 and removed in C++17 [1], so it was just
a matter of time when C++ compilers started to strictly enforce
the C++17 standard.

[1] https://en.cppreference.com/w/cpp/utility/functional/unary_function

Change-Id: Id97af5f7300b2ca2eef4edf33712fa6ec93daa96
Reviewed-on: http://gerrit.cloudera.org:8080/21196
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>