[gutil] fix compilation with clang 15
std::{binary,unary}_function have been deprecated since C++11 and
removed in C++17 [1], and it seems that CLANG 15 which comes with
Xcode 15.3 on macOS enforces this particular part of the C++17 standard.

In this instance, it's easy to address the issue because nothing
essential was used from std::unary_function and std::binary_function
but just typedefs which are no longer needed by the new standard library
anyway.  Since Kudu switched to C++17 a long time ago [2], the solution
is to stop inheriting from the deprecated classes.

[1] https://en.cppreference.com/w/cpp/utility/functional/unary_function
[2] https://github.com/apache/kudu/commit/e432e637e282ba87d41c02088a801f0838571edc

Change-Id: I3329427c3babd57418c83ab2401b7c4d29e14020
Reviewed-on: http://gerrit.cloudera.org:8080/21206
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>