[rpc] relax settings for MeasureAcceptorDispatchTimes
On some nodes, RpcAcceptorBench.MeasureAcceptorDispatchTimes might
run out of file descriptors if run with 16 concurrent client threads,
failing the test.  This patch modifies the default settings for the
scenario to run with just a single client thread by default.  That's
enough to provide good coverage for the related metrics while easing
the load on a test node.

Change-Id: Ic9af59046d6bcf28134cc88e7ff9e54643efc37f
Reviewed-on: http://gerrit.cloudera.org:8080/21212
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>