KUDU-613: Integrate SLRU cache into block cache
This patch integrates the new SLRU cache into the
block cache. It adds several flags to make the SLRU
cache configurable for capacity and lookups as well
as adding an eviction_policy flag to give users
the option to use the new SLRU cache. The default
value is still the old LRU cache.

Many integration tests that use the block cache
are also updated to include the block cache of
SLRU eviction policy.

Change-Id: I04411ab2756045f15a272f3397d46d871b087b03
Reviewed-on: http://gerrit.cloudera.org:8080/21390
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>