[c++17] fix unsorted compilation warnings with CLANG15
As a follow-up to a few prior patches, this patch addresses the rest
of warnings output by CLANG15 when compiling the Kudu project
on macOS Sonoma.

There are still some warnings from the linker, e.g. warnings about
the difference between target OS versions of the LLVM libraries
from 3rd-party and Kudu binaries being linked.  I think it's better
to address them in a separate patch, if ever doing so.

Change-Id: Iab094ec0766994d4e4d399a4ae1eb112ec16e5ee
Reviewed-on: http://gerrit.cloudera.org:8080/21242
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>