KUDU-3433 Fix flakiness in ClientTest.TestDeleteWithDeletedTableReserveSecondsWorks
The reason for the flakiness of the unit test is that we did not consider corner
cases, namely, the time reserved from detecting a table in soft-deleted state
until the execution of deletion, during which some operations may occur. We need
to reserve more time.

However, in order to verify that FLAGS_table_reserve_seconds is indeed effective,
the additional time cannot be too long. Taking these two factors into consideration,
I added one second of waiting time and performed multiple tests locally using the
gtest_repeat parameter to ensure that this modification is effective.

Change-Id: Iab3a6a5f54701470e368c043133f4646aa71126d
Reviewed-on: http://gerrit.cloudera.org:8080/21268
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>