[util] expose recently seen value in a histogram
When troubleshooting a performance issue with a Kudu cluster, I found it
would be useful if each of the existing histogram metrics exposed most
recently seen value (in particular, I was interested in knowing
the most recently captured length of a tablet's prepare queue).

This does not align with the semantics of a histogram since collected
statistics on the distribution of observed values aren't supposed to
include any notion of recency.  However, I think this is a useful
improvement from the observability standpoint because it allows for
collection of valuable information on monitored parameters
without introducing additional metrics.

NOTE: even if chromium-based Atomics are obsolete and STL atomics
      should be used in new code instead, I opted to use the former
      because it would look much uglier otherwise.  I think
      it's a better option to switch to using STL atomics in
      hdr_histogram.{h,cc} altogether in a separate changelist.

Change-Id: Ia4547faba050e09e31c83372105a9fe97b77ccbc
Reviewed-on: http://gerrit.cloudera.org:8080/21321
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Reviewed-by: Wang Xixu <1450306854@qq.com>
Tested-by: Alexey Serbin <alexey@apache.org>