Allow creating a table with empty range partition
Currently, when creating a table without setting range partition,
it will create an unbounded partition defaultly. In some cases,
this method is not suitable. For example, the user wants to create
a table with empty partition, and add a new partition daily to
import the data into Kudu. Besides, copying a empty partition table
to another cluster will create an unbounded partition table when
using the CLI command 'kudu table copy'. It changes the orginal
table structure.
Therefore, it is better to support to create a table with empty
partition, that is without any partitions.

Change-Id: I480072a693e5b6f1a13f86cb1475cff0452472e7
Reviewed-on: http://gerrit.cloudera.org:8080/21328
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>
Tested-by: Yingchun Lai <laiyingchun@apache.org>