KUDU-3566 fix summary metrics in Prometheus format
This patch corrects the output of various Kudu metrics backed by HDR
histograms.  From the Prometheus perspective, those metrics are output
as summaries [1], not histograms [2].  It's necessary to mark them
accordingly to avoid misinterpretation of the collected statistics.

I updated corresponding unit tests and verified that the updated output
was properly parsed and interpreted by a Prometheus 2.50.0 instance
running on my macOS laptop.

[1] https://prometheus.io/docs/concepts/metric_types/#summary
[2] https://prometheus.io/docs/concepts/metric_types/#histogram

Change-Id: I1375ddf1b0ecd730327cd44b4955813b80107f7b
Reviewed-on: http://gerrit.cloudera.org:8080/21338
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>