Minor refactoring on Op class
We have derived various types of operations based on the Op class.
Now, if we need to add time statistics for these operations, it
would be repetitive to add an initial timestamp separately for
each derived class.

In this patch, I moved the 'start_time_' from the WriteOp class to
the OpState class, making it easier for subsequent derived classes
to use.

Since there are no logical changes, no additional unit tests have
been added.

Change-Id: Ie391d4a55b8da08a62025a05cc466fc2b947099c
Reviewed-on: http://gerrit.cloudera.org:8080/21342
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>
Tested-by: Yingchun Lai <laiyingchun@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>