KUDU-3569 fix race in CFileSet::Iterator::OptimizePKPredicates()
This patch addresses data race reported in KUDU-3569, using
the tablet schema defined by the iterator's projection instead of
the schema stored in the tablet metadata file.  The latter might
be swapped by concurrently running AlterTable, which is the root
cause of the data race.

This is a follow-up to 936d7edc4e4b69d2e1f1dffc96760cb3fd57a934.

Change-Id: I92daa74cb86a77a4350f42db9ca5dec3a0d4ff75
Reviewed-on: http://gerrit.cloudera.org:8080/21359
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>