[build] Fix RocksDB Snappy dependency.
RocksDB uses Snappy. When statically linking Kudu, librocksdb.a contains
multiple undefined symbols that are defined in libsnappy.a. In the case
of static linking, the order of libraries passed to the linker matters,
and libsnappy.a must come after librocksdb.a. This sometimes causes the
release build to fail on RHEL 9.3.

Change-Id: I3ce75f69d94436f732dbe9a0011546b1ae494824
Reviewed-on: http://gerrit.cloudera.org:8080/21370
Reviewed-by: Attila Bukor <abukor@apache.org>
Tested-by: Attila Bukor <abukor@apache.org>