[tablet] update the severity level of some tablet metrics
The motivation for this update was a request to provide information
on the number of rows returned by scanners run against a particular
tablet. That's to be on par with the following tablet metrics of the
MetricLevel::kInfo severity level:
  * rows_inserted
  * rows_updated
  * rows_upserted
  * rows_deleted

I also took the liberty of updating the level of some other tablet
metrics to make their severity level more consistent with the rest
of tablet metrics that have similar semantics.

Change-Id: If468eeef8488f4a28bdc23a2fc45073e4a3592d4
Reviewed-on: http://gerrit.cloudera.org:8080/21384
Reviewed-by: Attila Bukor <abukor@apache.org>
Tested-by: Attila Bukor <abukor@apache.org>