[tserver] limit number of trace metrics for TabletCopyClient
The trace metrics registry assumes that the number of entries is quite
small, with the current threshold of 100.  In its turn, a thread pool
unconditionally registers and updates its trace metrics while executing
submitted tasks.  Since a tablet server might host thousands of tablet
replicas, it's not a good idea to include the UUID of the tablet into
the name of the download thread pool spawned by every TabletCopyClient
instance.

This is a follow-up to 0d95304fa46ee5d96bcaa934c7660369f2860e06.

Change-Id: I334aa81aaed2378e7cae558bd8bb9e0f0c970fec
Reviewed-on: http://gerrit.cloudera.org:8080/21393
Tested-by: Marton Greber <greber.mx@gmail.com>
Reviewed-by: Marton Greber <greber.mx@gmail.com>