[security-flags-itest] Fix missing command line flags
It is a known phenomenon, that static libraries won't be included into
an executable if there are no usage for any function or variable[1].
This means the initialization routines in the library won't be executed,
even if these initialization routines have side effects, such as
registering the variable in the gflags ecosystem. As a result, the
CheckRpcAuthnFlagsGroupValidator test failed because
"rpc_authentication" flag was not registered properly. To solve this
issue, a command line variable check is added, so now the library will
be used in the executable and the initialization routines will be
executed.

[1] https://stackoverflow.com/questions/1229430/how-do-i-prevent-my-unused-global-variables-being-compiled-out-of-my-static-li

Change-Id: Iec751e8761562612d97b886740c9b20cd134a0bc
Reviewed-on: http://gerrit.cloudera.org:8080/21399
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>