[glog] Fix minidump-test for cold symbols.
glog 0.6.0 cannot reliably resolve cold symbols. This was only fixed in
recent versions:
https://github.com/google/glog/issues/869

This might cause ASSERT_DEATH checks to fail in minidump-test. The long-
term solution would be to update to a newer version, but 0.7.1 has
a serious issue: stacktrace print crashes for long functions.
Since leaving cold symbols this way is not really an issue, backporting
the fixes just does not seem worth the time.

One affected platform is RHEL 9.3 with default g++.

Change-Id: I18d88790d09e9763223a767617794332859cbfad
Reviewed-on: http://gerrit.cloudera.org:8080/21385
Reviewed-by: Zoltan Chovan <zchovan@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>