[metrics] Add metrics for tablet replica election
Add tablet-level statistics to track the time consumed for
replica leader election. For any tablet that initiates a
leader election operation, the time will be recorded regardless
of the outcome.

The addition of monitoring items will aid in historical issue
tracking and analysis, as well as facilitate the configuration
of monitoring alarms.

Change-Id: I2d501b514e90461998bbc64f77f2bb12805ebfaa
Reviewed-on: http://gerrit.cloudera.org:8080/21490
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>
Reviewed-by: Zoltan Chovan <zchovan@cloudera.com>
Tested-by: Yingchun Lai <laiyingchun@apache.org>