KUDU-3584 fix flakiness in TableKeyRangeTest
When running client-test in TSAN/ASAN builds, the
TableKeyRangeTest.TestGetTableKeyRange scenario would sometimes fail
on busy nodes.  This patch addresses the issue by increasing the
timeout for scanners and for the write session, and making the related
code more robust overall.

I also took the liberty of cleaning up the related code.

Change-Id: I1efc2fe6ee7f2dfe94b52b14f1316ffbafd39d52
Reviewed-on: http://gerrit.cloudera.org:8080/21506
Reviewed-by: KeDeng <kdeng.ke@gmail.com>
Reviewed-by: Marton Greber <greber.mx@gmail.com>
Tested-by: Marton Greber <greber.mx@gmail.com>