[metrics] Add metrics for tablet replication time
Add tablet-level metric to track the time cost of
replication between replicas.

To verify the correctness of the new logic, I
constructed a synchronization scenario based on
write operations.

The addition of monitoring items will aid in
historical issue tracking and analysis, as well
as facilitate the configuration of monitoring alarms.

Change-Id: I4ea89c3954ceecabdb2998d2a8945f12a2fb8aa9
Reviewed-on: http://gerrit.cloudera.org:8080/21507
Reviewed-by: Zoltan Chovan <zchovan@cloudera.com>
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>
Tested-by: Yingchun Lai <laiyingchun@apache.org>