Fix cache cleaning in dense_node-itest
Running dense_node-itest with --measure_startup_drop_caches (as
src/kudu/scripts/benchmarks.sh does) fails on Ubuntu 22.04. We try to
open the /proc/sys/vm/drop_caches file with O_RDWR mode to write "3"
into it. On Ubuntu 18.04, the file seems to have read permissions too,
but on 22.04 it is properly set to --w-------, so we fail to do so.

Change-Id: I1f36f5a97d9a032aeb495989b4dc05191bf66425
Reviewed-on: http://gerrit.cloudera.org:8080/21509
Reviewed-by: Zoltan Chovan <zchovan@cloudera.com>
Reviewed-by: Wang Xixu <1450306854@qq.com>
Tested-by: Marton Greber <greber.mx@gmail.com>
Reviewed-by: Marton Greber <greber.mx@gmail.com>