[log] fix the Ranger client log count limit not working
In scenarios where authentication is enabled, the master
starts a Ranger subprocess to communicate with the KMS
process. I noticed that in these scenarios, the number
of log files generated by the Java subprocess was not
limited, causing a waste of disk space.

To address this issue, I updated the configuration file
for the Java subprocess. To verify the effectiveness of
this commit, I also added new unit tests to ensure it.

Change-Id: Idc528d68947c222fa7be338057ea7af134eb0dd4
Reviewed-on: http://gerrit.cloudera.org:8080/21572
Reviewed-by: Zoltan Chovan <zchovan@cloudera.com>
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>
Tested-by: Yingchun Lai <laiyingchun@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>