KUDU-3589 Wait until masters become voters
In SecurityITest.TestNonDefaultPrincipalMultipleMaster test, when the
leader was shut down, it was possible that some of the Raft members
are still in learner mode. This makes it impossible for the cluster to
recover and elect a new leader. To solve this test issue, there is an
explicit wait until masters become voters in the Raft consensus. This
was done by using the already existing VerifyVotersOnAllMasters
function. The function was moved from dynamic_multi_master-test to
external_mini_cluster to be usable from security-itest, too.

Change-Id: I1ea46345680ce506206e9335a147b887fc173b50
Reviewed-on: http://gerrit.cloudera.org:8080/21636
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>
(cherry picked from commit c5993dbf4f8e706784c690f3d5b38cad75276387)
Reviewed-on: http://gerrit.cloudera.org:8080/21996
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>